Move sfutue source to src/ from lib/ directory.
