+++ README.md
@@ -28,7 +28,7 @@ cd {version}
 ```
 make
 ```
-Now you have ```{path to sfuture}/lib/sfuture.js```. Use it.
+Now you have ```{path to sfuture}/src/sfuture.js```. Use it.
 
 ## How to contribute
 sfuture is written with TypeScript, checks style with tslint, runs tests with mocha and checks test coverage with istanbul. ```npm test``` command tests them all.
