+++ test/andThen.ts
@@ -1,5 +1,5 @@
 import assert = require('assert');
-import Future = require('../lib/future');
+import Future = require('../src/future');
 import should = require('./should');
 
 describe('#andThen', () => {
