+++ test/fallbackTo.ts
@@ -1,6 +1,6 @@
 import assert = require('assert');
 import should = require('./should');
-import Future = require('../lib/future');
+import Future = require('../src/future');
 
 describe('#fallbackTo', () => {
   it('does not affect on successful future.', (done: MochaDone) => {
