+++ test/sequence.ts
@@ -1,6 +1,6 @@
 import assert = require('assert');
 import should = require('./should');
-import Future = require('../lib/future');
+import Future = require('../src/future');
 
 describe('#sequence', () => {
   it('collects futures and returns a new future of their results.', (done: MochaDone) => {
