+++ test/successful.ts
@@ -1,6 +1,6 @@
 import assert = require('assert');
 import should = require('./should');
-import Future = require('../lib/future');
+import Future = require('../src/future');
 
 describe('#successful', () => {
   it('creates an already completed successful future with the specified result.', (done: MochaDone) => {
